requirements.txt
wheel
lxml
chromedriver
numpy
mapper
scipy
matplotlib
opencv-python-headless
scikit-image
argparse
imutils
python-telegram-bot==12.8
telethon
aiohttp
bleach
emoji
feedparser
future
geopy
googletrans
hurry.filesize
markdown2
nltk
pillow
psycopg2-binary
pynewtonmath
python-rapidjson
PyYAML
requests
spongemock
sqlalchemy
tldextract
wikipedia
zalgo-text
gTTs
gTTs-token
pyfiglet
beautifulsoup4
pytz
youtube.dl
google-api-python-client
pyDownload
telegraph
cowpy
psutil
hachoir
tswift
pySmartDL
pybase64
selenium
python-dotenv
git+https://github.com/rachmadaniHaryono/we-get.git
coffeehouse
html2text
spotdl
fortune
coffeehouse
python-barcode
PyDictionary
idna<2.9,>=2.5
fontTools
pymongo
dnspython
heroku3
gitpython
bing_image_downloader
alphabet_detector
Text2Emotion
howdoi
git+https://github.com/MissJuliaRobot/better_profanity.git
nudepy
